Fix reflow with interactive graphs
When 'playing' an interactive graph, there was a slight dimensions
change, causing a reflow in the page. This is because we need to match
the type of the img thumbnail and the canvas. Reset both to inline-block
and vertical-align middle, to avoid being caught with unexpected changes
in the dimensions of the content due to the content using inline
formatting model.g
http://www.w3.org/TR/CSS2/visuren.html#inline-formatting

Change-Id: I4cab574a61ea4807b9ee9174a37d2b0aa5d90fd5