Simplify Hooks section in extension.json
It is possible, but very rare to have more than a single hook handler
per hook in a single extension. The value can be an array or string in
both version 1 and 2 of extension.json.

This patch also removes a pointless comment that describes a constructor
as being a constructor.

Change-Id: I5d44df0d4e52705d08c1a52088e4293151497aa8