Drop ability to attempt server-side rendering with Graphoid
This functionality is no longer deployed, and was subject to
significant bitrot even before this.

Includes fixes for the old lazyload mode, in case we ever want
to reintroduce that behavior. Currently hard disabled with
the $alwaysMode variable in ParserTag.php, but required to make
sure we didn't break it any further with this change.

Bug: T242855
Change-Id: Ie032d4bab0fd1772eacd59f16da1e0d42ef303bc