Fix loading spinner when overflowing on mobile
The mw-graph element has a background with the spinner and the
overflow statement. The background of the div with the vega class
is white, the loaded canvas element graph is transparent. When
the canvas overflows, the vega div does NOT overflow, when you
scroll you thus get the background of the canvas directly
overlaying the background with the spinner.

Set the background on the canvas instead.

Bug: T286112
Change-Id: I8d0ca6d22677b1d30359664466675d98707ea9be