ve.dm.MWGraphModel: Fix invalid assert.throws() call in test
Follows-up I6282a08320c8624328, which introduced
the ve.dm.MWGraphModel#removeProperty method. In an earlier iteration
of the patch, it would throw an error when deleting unknown properties.
In the end, it was decided to silently ignore those.

The test was using `assert.throws()` incorrectly. Instead of calling
the method under test from a closure, it called it directly in the
test and passing its (undefined) return value to `throws()`.

These two mistakes cancelled each other out. If it was in a closure,
it would have failed the test because the method doesn't actually
throw. If it was as-is but actually threw, it would fail the test
for having an uncaught error in the test function.

Assuming the code is working correctly, fix the test to instead
assert that the object is unchanged, by repeating the previous
assertion.

This mistake was caught by QUnit 2.17, which improved validation
of the `assert.throws()` method. The reason the test passed previously
is that when you take the undefined value and store in a variable,
and then try to call `undef()`, then that does of course throw an
error.

This blocks Ia8ce57fa which upgrades QUnit in MediaWiki core.

Change-Id: Ia98a60166209646cc641d45a1ece62c131bcbcd0