Fix data collection of multiple graphs in Parsoid, take 2
The onParserAfterParse is called once by extension tag instance
in Parsoid context, which rewrites the jsConfigVars over and
over again, preventing display of multiple graphs on a page.

In this patch, we move the jsConfigVars writing to the
OutputPageParserOutput hook, and we fix the read/write access to the
extension data by making the keys contain the hash of the graph.
This way, the extension data is in write-only in the tags, and the
jsConfigVars are written only once at the end of the parse.

I44dde1d got reverted because it was missing the ApiGraph.php part, and
we wanted to avoid deploying an incomplete patch, which could have
happened considering the train timing.

Bug: T272942
Change-Id: I83a2d346970914304234c498a14e140f8415885e