Fix data collection of multiple graphs in Parsoid
The onParserAfterParse is called once by extension in Parsoid context,
which rewrites the jsConfigVars over and over again, preventing display
of multiple graphs on a page.

In this patch, we move the jsConfigVars writing to the
OutputPageParserOutput hook, and we fix the read/write access to the
extension data by making the keys contain the hash of the graph.
This way, the extension data is in write-only in the tags, and the
jsConfigVars are written only once at the end of the parse.

Bug: T272942
Change-Id: I44dde1d0f62b571f386356ed9468548e10f657f7