Introduce click to load with error handling for unsupported graphs
This patch changes the way we load graphs requiring the user to click
the graph to load it (for modern browsers they will also load
automatically when scrolled into view via IntersectionObserver)

Before loading it we check we support it and show feedback to the user
if we don't.

Show error message if there is a problem loading.

Additional changes:
* Switch to ES6 linting given we no longer support ES5.
* As part of this change, all graphs will default to 500x500
if no dimensions are specified.

Bug: T73107
Bug: T272530
Change-Id: I3465888bb5a3475e82b7a0899bba151c4b029f23